+++ server/package.json
@@ -19,7 +19,6 @@
   "homepage": "https://github.com/lifegrain/fancy-todo#readme",
   "dependencies": {
     "bcrypt": "^4.0.1",
-    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "http-errors": "^1.7.3",
     "jsonwebtoken": "^8.5.1",
@@ -27,5 +26,8 @@
     "pg": "^7.18.2",
     "sequelize": "^5.21.5",
     "sequelize-cli": "^5.5.1"
+  },
+  "devDependencies": {
+    "dotenv": "^8.2.0"
   }
 }
